created config.php, script.js. refactoring code. #1 #hacktoberfest want
