+++ .gitignore
@@ -42,4 +42,5 @@ Network Trash Folder
 Temporary Items
 .apdisk
 
-/sma/vendor/
+/vendor/
+.idea/
