+++ Facebook/FacebookRedirectLoginHelper.php
@@ -183,8 +183,8 @@ class FacebookRedirectLoginHelper
         '/oauth/access_token',
         $params
       ))->execute()->getResponse();
-      if (isset($response['access_token'])) {
-        return new FacebookSession($response['access_token']);
+      if (isset($response->access_token)) {
+        return new FacebookSession($response->access_token);
       }
     }
     return null;
